requirement.txt
openai
python-dotenv
slack_bolt
slack_sdk